chore: fix analyzer warnings - remove unused locals, dead code and stackalloc leak
- CisoReader: remove unused outBuf
- CisoWriter: remove unused name in PackFromDirectoryForCiso
- RemapFilesystem: remove unused mappings vars, discard WaxGlob constructions and AvlInsert result
- XisoOperationsTests: remove unused PopulateSingle
- XisoFileTimeTests: remove unused srcDir out param from CreateSimpleIso
- XgdTablesTests: remove unused pvd14 local
- XisoReader: move stackalloc headerRest out of AuditWalk loop to avoid per-iteration allocation leak
- Additional style fixes across BattleTests/Cli/Core/Tests (IDE warnings)

+++ XISOSharp.Core/FileTimeHelper.cs
@@ -101,7 +101,7 @@ public static class FileTimeHelper
     {
         if (fileTime == 0UL)
             return FileTimeEpoch;
-        if (fileTime <= (ulong)long.MaxValue)
+        if (fileTime <= long.MaxValue)
         {
             try
             {
